+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # CDP Python SDK Changelog
 
-## [0.14.0] - 2025-01-16
+## Unreleased
 
 ### Added
 - Add `Webhook.delete_webhook` instance method to delete a webhook instance.
+- Add `skip_batching` option to `Wallet.transfer` to allow for lower latency gasless transfers.
 
 ### Deprecated
 - Deprecate `Webhook.delete` static method in `Webhook` which deletes a webhook by its ID.